requirements.txt
atomicwrites==1.2.1
attrs==18.2.0
factory-boy==2.11.1
Faker==0.9.2
inflection==0.3.1
more-itertools==4.3.0
pluggy==0.8.0
py==1.7.0
pytest==3.10.0
pytest-factoryboy==2.0.2
python-dateutil==2.7.5
six==1.11.0
text-unidecode==1.2
